Renaming some files, adding dependencies for testing
