+++ app/build.gradle
@@ -88,7 +88,17 @@ dependencies {
     kapt "com.google.dagger:dagger-android-processor:2.35.1"
     kapt "org.jetbrains.kotlinx:kotlinx-metadata-jvm:0.5.0"
 
+    //Local unit tests
     testImplementation 'junit:junit:4.13.2'
+    testImplementation "androidx.test:core-ktx:1.5.0"
+    testImplementation "androidx.test.ext:junit-ktx:1.1.5"
+    testImplementation "androidx.test:runner:1.5.2"
+    testImplementation "androidx.arch.core:core-testing:2.2.0"
+    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6.4"
+    testImplementation 'org.robolectric:robolectric:4.7.3'
+
+    //Local Instrumented test
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'androidx.test:runner:1.5.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 }
