+++ app/src/main/java/com/globant/imdb/data/MovieFromRemote.kt
@@ -1,9 +1,9 @@
 package com.globant.imdb.data
 
-import com.globant.imdb.database.MovieInDB
+import com.globant.imdb.database.Movie
 import com.google.gson.annotations.SerializedName
 
-data class Movie(
+data class MovieFromRemote(
     val adult: Boolean,
     @SerializedName("backdrop_path") val backdropPath: String,
     @SerializedName("genre_ids") val genreIds: List<Int>,
@@ -20,9 +20,9 @@ data class Movie(
     @SerializedName("vote_count") val vote_count: Int
 )
 
-fun List<Movie?>.asDBModel(): List<MovieInDB> {
+fun List<MovieFromRemote?>.asDBModel(): List<Movie> {
     return mapNotNull {
-        MovieInDB(
+        Movie(
             posterPath = it?.posterPath ?: "",
             title = it?.title ?: "",
             releaseDate = it?.releaseDate ?: "",
