+++ app/src/main/java/com/globant/imdb/data/TopRatedMovies.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class TopRatedMovies(
     val page: Int,
-    val results: List<Movie>,
+    val results: List<MovieFromRemote>,
     @SerializedName("total_pages") val totalPages: Int,
     @SerializedName("total_results") val totalResults: Int
 )
