+++ app/src/main/java/com/globant/imdb/repo/MoviesRepo.kt
@@ -3,17 +3,17 @@ package com.globant.imdb.repo
 import android.util.Log
 import com.globant.imdb.api.IMDbNetworking
 import com.globant.imdb.data.Constants.API_KEY
-import com.globant.imdb.data.Movie
+import com.globant.imdb.data.MovieFromRemote
 import com.globant.imdb.data.asDBModel
-import com.globant.imdb.database.MovieInDB
+import com.globant.imdb.database.Movie
 import javax.inject.Inject
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 
 class MoviesRepo @Inject constructor() {
 
-    suspend fun getMoviesByQuery(query: String): List<MovieInDB> {
-        var movieListByQuery: List<Movie>? = null
+    suspend fun getMoviesByQuery(query: String): List<Movie> {
+        var movieListByQuery: List<MovieFromRemote>? = null
         withContext(Dispatchers.IO) {
             try {
                 val response = IMDbNetworking.movieData.getMoviesFromQueryAsync(
