+++ app/src/test/java/com/globant/imdb/localDB/IMDbDaoTest.kt
@@ -1,0 +1,4 @@
+package com.globant.imdb.localDB
+
+class IMDbDaoTest {
+}
